UI/requirements.txt
customtkinter
matplotlib
numpy
mido
python-rtmidi
pyserial